refactor(Hearings): moved `deleteHearing` logic from Controller to Service Implementation

+++ src/main/java/com/intelliseven/spring_data_jpa_activity/service/impl/HearingServiceImpl.java
@@ -82,7 +82,11 @@ public class HearingServiceImpl implements HearingService {
   }
 
   @Override
-  public void delete(Long id) {
+  public ResponseEntity<?> deleteHearing(Long id) {
+    if (!hearingRepo.existsById(id)) {
+      return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+    }
     hearingRepo.deleteById(id);
+    return new ResponseEntity<>(HttpStatus.NO_CONTENT);
   }
 }
